master: billing basics

+++ app/src/main/java/mx/alxr/voicenotes/feature/recognizer/TranscriptionArgs.kt
@@ -3,5 +3,5 @@ package mx.alxr.voicenotes.feature.recognizer
 import mx.alxr.voicenotes.repository.record.RecordEntity
 
 data class TranscriptionArgs(val entity: RecordEntity? = null,
-                             val requiredCoins:Int = 0,
-                             val availableCoins:Int = -1)
+                             val requiredCoins:Long = 0,
+                             val availableCoins:Long = -1)
